app/requirements.txt
Flask
Flask_RESTful
mysql-connector-python
Jinja2